Move scripts to the corresponding folders
