+++ Assets/Scripts/Background/Parallax.cs.meta
