+++ Assets/Scripts/Enemy/EnemyWalkerMovement.cs
@@ -1,4 +1,5 @@
 ﻿using UnityEngine;
+using Wall;
 
 namespace Enemy
 {
