+++ Assets/Scripts/HandleObjectFacing.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using UnityEngine;
+using Wall;
 
 public class HandleObjectFacing : MonoBehaviour
 {
