+++ Assets/Scripts/Player/PlayerWallCollisions.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Resources;
 using UnityEngine;
+using Wall;
 
 namespace Player
 {
