+++ Assets/Scripts/Spawn/SpawnOil.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using Player;
 using UnityEngine;
+using Wall;
 
 namespace Spawn
 {
